simplify serializablevalue definition and move test to runner.test.ts
